Improved tests (#37)
* new action to force test coverage threshold before PRs and merging to main

* all image and string assertions have been fixed in tests, no memory leaks

* all tests pass again but that reintroduces some image and string assertion issues, major 'MVC' refactor likely necessary to actually address the issues. out of scope for now

* more Rack tests (currently failing)

* small fix to attempt to improve save corruption

* adding logging around DAW corrution to test in Cubase

* fixing some initalValue issues and changing double click to cmd+click for better UX

* graceful failure when saving while UI is not open

Merging so that all code is in main and can be used to create the rewrite branch. The build DOES work locally and the plugin runs as expected (as a user). Tests and Coverage may or may not pass but doesnt matter since the rewrite would break it all ayway.